Extract FileBatchIterator from filter
